fix: Split tests for runtime.
