+++ .github/workflows/ci.yml
@@ -352,6 +352,27 @@ jobs:
       - name: Run runtime package test suite
         run: cd packages/control && pnpm test
 
+  ci-runtime-lint:
+    needs: setup-node_modules
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    steps:
+      - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
+      - uses: pnpm/action-setup@v4.0.0
+      - uses: actions/setup-node@v4
+        with:
+          node-version: 22
+          cache: 'pnpm'
+      - name: pnpm install
+        uses: nick-fields/retry@v3.0.0
+        with:
+          max_attempts: 10
+          timeout_minutes: 15
+          retry_on: error
+          command: pnpm install --frozen-lockfile
+      - name: Run runtime package test suite
+        run: cd packages/runtime && pnpm run lint
+
   ci-runtime:
     needs: setup-node_modules
     runs-on: ${{matrix.os}}
@@ -360,6 +381,7 @@ jobs:
       matrix:
         node-version: [20, 22]
         os: [ubuntu-latest, windows-latest]
+        suite: [main, api, cli, start, types, workers]
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4
       - uses: pnpm/action-setup@v4.0.0
@@ -375,7 +397,7 @@ jobs:
           retry_on: error
           command: pnpm install --frozen-lockfile
       - name: Run runtime package test suite
-        run: cd packages/runtime && pnpm test
+        run: cd packages/runtime && pnpm run test:${{ matrix.suite }}
 
   ci-service:
     needs: setup-node_modules
