+++ packages/runtime/config.d.ts
@@ -26,7 +26,7 @@ export type HttpsSchemasPlatformaticDevPlatformaticRuntime270Json = {
   services?: {
     [k: string]: unknown;
   }[];
-  workers?: number;
+  workers?: number | string;
   web?: {
     [k: string]: unknown;
   }[];
