+++ packages/runtime/schema.json
@@ -93,15 +93,29 @@
             "type": "boolean"
           },
           "workers": {
-            "type": "number",
-            "minimum": 1
+            "anyOf": [
+              {
+                "type": "number",
+                "minimum": 1
+              },
+              {
+                "type": "string"
+              }
+            ]
           }
         }
       }
     },
     "workers": {
-      "type": "number",
-      "minimum": 1,
+      "anyOf": [
+        {
+          "type": "number",
+          "minimum": 1
+        },
+        {
+          "type": "string"
+        }
+      ],
       "default": 1
     },
     "web": {
@@ -140,8 +154,15 @@
             "type": "boolean"
           },
           "workers": {
-            "type": "number",
-            "minimum": 1
+            "anyOf": [
+              {
+                "type": "number",
+                "minimum": 1
+              },
+              {
+                "type": "string"
+              }
+            ]
           }
         }
       }
